refactor: extract overlay subsystem from windows.rs to satisfy file size gate
platform/windows.rs was 1535 lines (max 1500 per AGENTS.md file size
policy). Extract the self-contained overlay subsystem into a dedicated
windows/overlay.rs module (531 lines), bringing the parent module to
1049 lines.

Changes:
- New: windows/overlay.rs — OverlayLayer, ControlOverlay, OverlayTone,
  and all overlay creation/rendering functions
- windows.rs: remove extracted code, add mod overlay + re-exports
- geometry.rs: widen 7 functions from pub(super) to pub(crate) so the
  sibling overlay module can access them
- RAII guards (RegisteredHotKey, RegisteredSessionNotifications,
  RegisteredDesktopBarrier) remain in windows.rs

Build: cargo check + cargo test pass with zero warnings.
